fix chicken and egg
